requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
openai
python-dotenv